backend/pyproject.toml
[tool.poetry]
name = "kpi-backend"
version = "0.1.0"
description = "KPI-Inteligence techincal test"
authors = ["VD"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.7"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
coverage = "^5.5"
django-stubs = "^1.7.0"
factory_boy = "^3.2.0"
faker = "^6.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
